Fix Create and Delete callbacks.
Create Callback is made synchronous, the async was somehow very much
broken. It caused duplicate items to be included and did not work with
the new bootstrap HTML layout.
The "variants" and "stock" tabs are sync too. So this brings it more in
line with the other tabs.

Delete now simply replaces the table with the new table.